LEKSE-92/Fix service index
