+++ src/grades.ts
@@ -1,0 +1,10 @@
+const GRADES = [
+  { gradeID: '8', label: '8. klasse' },
+  { gradeID: '9', label: '9. klasse' },
+  { gradeID: '10', label: '10. klasse' },
+  { gradeID: '11', label: 'Vg 1' },
+  { gradeID: '12', label: 'Vg 2' },
+  { gradeID: '13', label: 'Vg 3' },
+];
+
+export default GRADES;
