+++ src/services/util-service.ts
@@ -1,0 +1,7 @@
+import GRADES from '../grades';
+
+export const studentGradeFormat = (studentGrade: string): string => {
+  return GRADES.filter(({ gradeID }) => gradeID === studentGrade).map(
+    ({ label }) => label,
+  )[0];
+};
