+++ src/styles/admin-questions.less
@@ -11,7 +11,7 @@
     display: grid;
     grid-template-rows: auto;
     grid-row-gap: 1rem;
-
+    margin-top: 1rem;
     &-item {
       border: thin solid;
       padding: 0.5rem;
@@ -39,7 +39,7 @@
   margin-bottom: 1rem;
   width: 14rem;
   p {
-    color: @disabled-color;
+    color: @help-text-color;
   }
 }
